refactor: log 타입 따로 분리
