+++ frontend/src/apis/admin/log.ts
@@ -1,18 +1,19 @@
-import axios, { AxiosError, type AxiosResponse } from 'axios';
-import type { IResultType, IGetUserLog } from '@/types/api';
+import { AxiosError, type AxiosResponse } from 'axios';
+import type { IResultType } from '@/types/api';
 import { ApiType } from '@/constants/apiType';
 import { apiInstance } from '..';
+import type { IGetUserLog } from '@/types/log';
 
 const api = apiInstance();
 
 const manageLogAPI = {
   endPoint: {
-    userlog: 'api/admin/user-log/record'
+    userLog: 'api/admin/user-log/record'
   },
   headers: {},
   selectLog: (): Promise<IResultType<Array<IGetUserLog>>> => {
     return api
-      .get(manageLogAPI.endPoint.userlog)
+      .get(manageLogAPI.endPoint.userLog)
       .then((res: AxiosResponse) => {
         const data = res.data;
         return { isSuccess: true, data: data.userLogRecord, type: ApiType.ADMIN_LOG };
