+++ frontend/src/types/log.ts
@@ -1,0 +1,6 @@
+export interface IGetUserLog {
+  email: string;
+  nickname: string;
+  lastModifiedAt: string;
+  banCount: number;
+}
